remove directive
