+++ demo/index.ts
@@ -38,7 +38,6 @@ class HeroSearchService {
 
 class ValueType {}
 
-
 interface Props {
   a: number;
 }
@@ -54,7 +53,7 @@ interface Props {
   `),
 })
 
-class RouteChild implements OnInit, HasRender, ReceiveProps {
+class RouteChild implements OnInit, HasRender, ReceiveProps, OnDestory {
   public setState: SetState;
   public state: any;
   public heroSearchService: HeroSearchService2;
@@ -103,6 +102,10 @@ class RouteChild implements OnInit, HasRender, ReceiveProps {
     //   b: nextProps.a,
     // });
   }
+
+  public nvOnDestory() {
+    console.log('RouteChild nvOnDestory');
+  }
 }
 
 @Component({
@@ -115,7 +118,7 @@ class RouteChild implements OnInit, HasRender, ReceiveProps {
     </div>
   `),
 })
-class PCChild implements OnInit, BeforeMount, AfterMount, ReceiveProps {
+class PCChild implements OnInit, BeforeMount, AfterMount, ReceiveProps, OnDestory {
   public props: any;
   public state: any;
   public setState: SetState;
@@ -165,6 +168,10 @@ class PCChild implements OnInit, BeforeMount, AfterMount, ReceiveProps {
     console.log('PCChild nvAfterMount props11', this.props.ax);
   }
 
+  public nvOnDestory() {
+    console.log('PCChild nvOnDestory');
+  }
+
   public nvReceiveProps(nextProps: any) {
     console.log(this.props.ax);
     console.log(4444, nextProps);
@@ -187,7 +194,7 @@ class PCChild implements OnInit, BeforeMount, AfterMount, ReceiveProps {
     </div>
   `),
 })
-class PComponent implements OnInit, WatchState, BeforeMount, AfterMount, ReceiveProps {
+class PComponent implements OnInit, WatchState, BeforeMount, AfterMount, ReceiveProps, OnDestory {
   public setState: SetState;
   public state: any;
   public props: any;
@@ -249,6 +256,9 @@ class PComponent implements OnInit, WatchState, BeforeMount, AfterMount, Receive
     console.log(1111111111111, nextProps);
     this.state.ax = nextProps.ax;
   }
+  public nvOnDestory() {
+    console.log('PComponent is nvOnDestory');
+  }
 }
 
 @Injected
@@ -270,7 +280,7 @@ class PComponent implements OnInit, WatchState, BeforeMount, AfterMount, Receive
     </div>
     `),
 })
-class R1 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange {
+class R1 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange, OnDestory {
   public hSr: HeroSearchService;
   public getLocation: GetLocation;
   public setLocation: SetLocation;
@@ -344,6 +354,10 @@ class R1 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange {
     this.setState({ a: a });
     // this.state.a = a;
   }
+
+  public nvOnDestory() {
+    console.log(this.getLocation(), 'R1 is nvOnDestory');
+  }
 }
 
 @Injected
@@ -361,7 +375,7 @@ class R1 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange {
     </div>
   `),
 })
-class R2 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange {
+class R2 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange, OnDestory {
   public getLocation: GetLocation;
   public setLocation: SetLocation;
   public state: any;
@@ -392,6 +406,11 @@ class R2 implements OnInit, BeforeMount, AfterMount, WatchState, RouteChange {
   public nvWatchState(oldState: any) {
     console.log('oldState Controller:', oldState);
   }
+
+  public nvOnDestory() {
+    console.log(this.getLocation(), 'R2 is nvOnDestory');
+  }
+
   public showAlert() {
     console.log('this.state.a', this.state.a);
     // alert('我错了 点下控制台看看吧');
@@ -442,7 +461,7 @@ class TestComponent implements OnInit, OnDestory, ReceiveProps {
   selector: 'container-wrap',
   template: (`
     <div>
-      <p nv-id="@countState($.a)" nv-if="@countState($.a)" nv-on:click="@changeInput()">{{$.a}}</p>
+      <p test-directive="$.a" nv-id="@countState($.a)" nv-if="@countState($.a)" nv-on:click="@changeInput()">{{$.a}}</p>
       <test-component nv-repeat="let man in $.testArray" nv-key="man.name" man="{@countState(man.name)}" nv-if="$.a"></test-component>
       <p nv-on:click="@go()">container: {{@countState($.a)}}</p>
       <input nv-model="$.a" />
@@ -478,8 +497,8 @@ class Container implements OnInit, AfterMount, WatchState {
     private value: ValueType,
   ) {
     this.hss.test();
-    console.log('nvHttp', nvHttp);
-    console.log('value', value);
+    console.log('nvHttp', this.nvHttp);
+    console.log('value', this.value);
     this.state = {
       a: 1,
       b: 3,
