+++ src/index.ts
@@ -1,18 +1,11 @@
 import 'core-js/modules/es6.array.find';
 import 'core-js/modules/es6.array.find-index';
-import 'core-js/modules/es6.array.for-each';
 import 'core-js/modules/es6.array.from';
-import 'core-js/modules/es6.array.index-of';
-import 'core-js/modules/es6.array.map';
-import 'core-js/modules/es6.array.sort';
-import 'core-js/modules/es6.object.define-property';
-import 'core-js/modules/es6.object.to-string';
+import 'core-js/modules/es6.function.bind';
+import 'core-js/modules/es6.reflect.construct';
 import 'core-js/modules/es6.regexp.match';
 import 'core-js/modules/es6.regexp.replace';
 import 'core-js/modules/es6.map';
-import 'core-js/modules/es6.promise';
-import 'core-js/modules/es6.function.bind';
-import 'core-js/modules/es6.reflect.construct';
 import 'core-js/modules/es6.promise';
 import 'reflect-metadata';
